+++ runtime/events.go
@@ -49,9 +49,3 @@ func (e Events) EmitKV(eventType string, attrs ...event.Attribute) error {
 	e.EventManagerI.EmitEvents(sdk.Events{sdk.NewEvent(eventType, attributes...)})
 	return nil
 }
-
-// EmitNonConsensus emits an typed event that is defined in the protobuf file.
-// In the future these events will be added to consensus.
-func (e Events) EmitNonConsensus(event protoiface.MessageV1) error {
-	return e.EventManagerI.EmitTypedEvent(event)
-}