+++ x/staking/keeper/abci.go
@@ -4,11 +4,10 @@ import (
 	"context"
 	"time"
 
-	abci "github.com/cometbft/cometbft/abci/types"
-
 	"cosmossdk.io/x/staking/types"
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
+	"github.com/cosmos/cosmos-sdk/types/module"
 )
 
 // BeginBlocker will persist the current header and validator set as a historical entry
@@ -19,7 +18,7 @@ func (k *Keeper) BeginBlocker(ctx context.Context) error {
 }
 
 // EndBlocker called at every block, update validator set
-func (k *Keeper) EndBlocker(ctx context.Context) ([]abci.ValidatorUpdate, error) {
+func (k *Keeper) EndBlocker(ctx context.Context) ([]module.ValidatorUpdate, error) {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyEndBlocker)
 	return k.BlockValidatorUpdates(ctx)
 }